Data storage class with SQLite provider

Implementation of the DataStorage R6 class to SQLite backend using a unified API for read/write operations

Super class

shiny.telemetry::DataStorage -> DataStoragePlumber

Active bindings

action_read_endpoint

string field that returns read action endpoint

session_read_endpoint

string field that returns read session endpoint

action_insert_endpoint

string field that returns insert action endpoint

session_insert_endpoint

string field that returns insert session endpoint

Methods

Public methods

DataStoragePlumber$new()

DataStoragePlumber$insert()

DataStoragePlumber$read_user_data()

DataStoragePlumber$read_session_data()

DataStoragePlumber$close()

DataStoragePlumber$clone()

Method new()

Initialize the data storage class

Usage

DataStoragePlumber$new(
  hostname = "127.0.0.1",
  port = 80,
  protocol = "http",
  path = NULL,
  secret = NULL,
  authorization = NULL
)

Arguments

hostname

string with hostname of plumber instance,

port

numeric value with port number of plumber instance.

protocol

string with protocol of the connection of the plumber instance.

path

string with sub-path of plumber deployment.

secret

string with secret to sign communication with plumber (can be NULL for disabling communication signing).

authorization

string to use in HTTP headers for authorization (for example: to authenticate to a plumber deployment behind a connect server).

Method insert()

Insert new data

Usage

DataStoragePlumber$insert(
  values,
  insert_time = TRUE,
  bucket = private$log_file_path
)

Arguments

values

list of values to write to database

insert_time

boolean flag that indicates if `time` parameters should be added automatically

bucket

name of table to write

Method read_user_data()

read all user data from SQLite

Usage

DataStoragePlumber$read_user_data(date_from, date_to)

Arguments

date_from

date representing the starting day of results

date_to

date representing the last day of results

Method read_session_data()

read all session data from SQLite

Usage

DataStoragePlumber$read_session_data(date_from, date_to)

Arguments

date_from

date representing the starting day of results

date_to

date representing the last day of results

Method close()

read all session data

Usage

DataStoragePlumber$close()

Method clone()

The objects of this class are cloneable with this method.

Usage

DataStoragePlumber$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) {
# Make sure the PLUMBER_SECRET environment variable is valid before
# running these examples (NULL or a valid secret)

data_storage <- DataStoragePlumber$new(
  hostname = "connect.appsilon.com",
  path = "shiny_telemetry_plumber",
  port = 443,
  protocol = "https",
  authorization = Sys.getenv("CONNECT_AUTHORIZATION_KEY"),
  secret = Sys.getenv("PLUMBER_SECRET")
)

data_storage <- DataStoragePlumber$new(
  hostname = "127.0.0.1",
  path = NULL,
  port = 8087,
  protocol = "http",
  secret = Sys.getenv("PLUMBER_SECRET")
)

telemetry <- Telemetry$new(data_storage = data_storage)
telemetry$log_login()

telemetry$log_click("an_id")
telemetry$log_click("a_different_id")

telemetry$log_session(detail = "some detail")

data_storage$read_user_data("2020-01-01", "2025-01-01")
data_storage$read_session_data("2020-01-01", "2025-01-01")
}